+++ packages/mini-editor/src/index.tsx
@@ -1,14 +1,14 @@
 import "./index.scss"
 
-import { MiniEditor, MiniEditorProps } from "./mini-editor"
 import {
-  MiniEditorWithState,
-  StatefulEditorProps,
-} from "./stateful-editor"
+  MiniEditorHike,
+  MiniEditorHikeProps,
+} from "./mini-editor-hike"
+import { MiniEditor, MiniEditorProps } from "./mini-editor"
 
 export {
+  MiniEditorHike,
   MiniEditor,
-  MiniEditorWithState,
+  MiniEditorHikeProps,
   MiniEditorProps,
-  StatefulEditorProps,
 }
